requirements.txt
airflow-dbt
dbt-bigquery